# public/openapi.yaml
openapi: 3.1.0
info:
  title: DataMaster SSO Authentication API
  version: 1.0.0
  description: |
    # DataMaster SSO Flow - OAuth 2.0 + PKCE

    **日本語**: このAPIは、DataMasterアプリケーションとSmart iMATE IDプロバイダー間のOAuth 2.0 + PKCE認証フローを実装します。AWS Cognitoをバックエンドとして使用し、マルチテナント対応のセキュアな認証を提供します。

    **Tiếng Việt**: API này triển khai luồng xác thực OAuth 2.0 + PKCE giữa ứng dụng DataMaster và nhà cung cấp định danh Smart iMATE. Sử dụng AWS Cognito làm backend và cung cấp xác thực an toàn hỗ trợ đa tenant.

    ## Security Features / セキュリティ機能 / Tính năng bảo mật

    - **PKCE (RFC 7636)**: Prevents authorization code interception attacks
      - **日本語**: 認可コード傍受攻撃を防止
      - **Tiếng Việt**: Ngăn chặn tấn công đánh cắp mã ủy quyền

    - **State Parameter**: CSRF protection via signed JWT
      - **日本語**: 署名付きJWTによるCSRF保護
      - **Tiếng Việt**: Bảo vệ CSRF qua JWT đã ký

    - **Multi-Tenant Isolation**: Tenant-scoped cookies and validation
      - **日本語**: テナント単位のCookieと検証
      - **Tiếng Việt**: Cookie và xác thực theo phạm vi tenant

    ## Architecture / アーキテクチャ / Kiến trúc

    ```
    DataMaster (Client)
           ↕ OAuth 2.0 + PKCE
    Smart iMATE (Identity Provider)
           ↕ USER_PASSWORD_AUTH
    AWS Cognito (User Pool)
    ```
servers:
  - url: http://localhost:3000
    description: |
      **日本語**: DataMaster開発環境
      **Tiếng Việt**: Môi trường phát triển DataMaster
  - url: http://localhost:8080
    description: |
      **日本語**: Smart iMATE IDプロバイダー
...
paths:
...
  /api/auth/callback:
    get:
      tags: [Phase 4]
      summary: Token Exchange - Authorization Code for Tokens
      description: |
        **日本語説明**:
        DataMasterがSmart iMATEから受け取った認可コードをアクセストークンと交換します。このフェーズではPKCE検証、state検証、client認証が行われます。

        **処理ステップ**:
        1. **State検証**: URLのstateとCookieのstateを比較（CSRF対策）
        2. **Code verifier取得**: Cookieから`oauth_code_verifier_{tenant}`を取得
        3. **Tenant config解決**: データベースからclient_id、client_secretなどを取得
        4. **Token exchange**: Smart iMATE `/oauth2/token`エンドポイントへPOSTリクエスト
        5. **PKCE検証**: Smart iMATEがSHA256(code_verifier) === stored_code_challengeを確認
        6. **Tokenレスポンス**: access_token、id_token、refresh_tokenを受信

        **セキュリティ検証**:
        - Authorization code JWT署名検証（RS256）
        - Code有効期限検証（10分）
        - PKCE検証: `SHA256(code_verifier) === code_challenge`
        - Client認証: client_id + client_secret
        - Redirect URI完全一致
        - 認可コード使用済みチェック（ブラックリスト）

        **Giải thích tiếng Việt**:
        DataMaster đổi mã ủy quyền nhận từ Smart iMATE lấy access token. Giai đoạn này thực hiện xác thực PKCE, state và client.

        **Các bước xử lý**:
        1. **Xác thực State**: So sánh state trong URL với Cookie (chống CSRF)
        2. **Lấy Code verifier**: Lấy `oauth_code_verifier_{tenant}` từ Cookie
        3. **Giải quyết Tenant config**: Lấy client_id, client_secret từ database
        4. **Đổi token**: POST request đến endpoint `/oauth2/token` Smart iMATE
        5. **Xác thực PKCE**: Smart iMATE kiểm tra SHA256(code_verifier) === stored_code_challenge
        6. **Token response**: Nhận access_token, id_token, refresh_token

        **Xác thực bảo mật**:
        - Xác thực chữ ký JWT mã ủy quyền (RS256)
        - Kiểm tra hết hạn code (10 phút)
        - Xác thực PKCE: `SHA256(code_verifier) === code_challenge`
        - Xác thực Client: client_id + client_secret
        - Khớp hoàn toàn Redirect URI
        - Kiểm tra mã ủy quyền đã sử dụng (blacklist)

      parameters:
        - name: code
          in: query
          required: true
          description: |
            **日本語**: 認可コードJWT（Smart iMATEから発行）
            **Tiếng Việt**: JWT mã ủy quyền (do Smart iMATE phát hành)
          schema:
            type: string
            example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...

        - name: state
          in: query
          required: true
          description: |
            **日本語**: 署名付きstate JWT（検証用）
            **Tiếng Việt**: JWT state đã ký (để xác thực)
          schema:
            type: string

      responses:
        '302':
          description: |
            **日本語**: Token取得成功、ダッシュボードへリダイレクト
            **Tiếng Việt**: Lấy token thành công, chuyển hướng đến dashboard
          headers:
            Location:
              schema:
                type: string
                example: /dashboard
            Set-Cookie:
              description: |
                **日本語**: セッションJWT Cookie（30日有効）
                **Tiếng Việt**: Session JWT Cookie (hiệu lực 30 ngày)
              schema:
                type: string
                example: session=eyJhbGc...; Max-Age=2592000; HttpOnly; Secure; SameSite=Lax; Path=/

        '400':
          description: |
            **日本語**: 無効なリクエスト（state不一致、code無効など）
            **Tiếng Việt**: Yêu cầu không hợp lệ (state không khớp, code không hợp lệ)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: |
            **日本語**: エラーコード
            **Tiếng Việt**: Mã lỗi
          example: invalid_grant
        error_description:
          type: string
          description: |
            **日本語**: エラー詳細
            **Tiếng Việt**: Mô tả chi tiết lỗi
          example: PKCE verification failed
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **日本語**: Cognito Access Token JWT（1時間有効）
        **Tiếng Việt**: JWT Cognito Access Token (hiệu lực 1 giờ)
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        **日本語**: セッションJWT Cookie（HS256署名、30日有効）
        **Tiếng Việt**: Session JWT Cookie (chữ ký HS256, hiệu lực 30 ngày)